+++ lib/globalVars.js
@@ -1,0 +1,9 @@
+const $title = $('#title-input');
+const $body = $('#body-input');
+const $userSearch = $('#search-box');
+const $h2 = $('h2');
+const $p = $('p');
+const importanceLevel = ["None", "Low", "Normal", "High", "Critical"];
+let topTen = true;
+let completedDisplayed = false;
+let showCompleteBtnToggle = false;
